Refator code into the correct locations
